+++ .github/workflows/coverage.yml
@@ -32,7 +32,7 @@ jobs:
     
     - name: Generate coverage report
       run: |
-        python -m pytest --cov=ap_cull_lights --cov-report=term --cov-report=xml || true
+        python -m pytest --cov=ap_cull_light --cov-report=term --cov-report=xml || true
         python -m coverage report | tail -1 | awk '{print $$4}' | sed 's/%//' > .coverage-percentage || echo "0" > .coverage-percentage
     
     - name: Extract coverage data and set badge info
