+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include LICENSE
 include README.md
-recursive-include ap_cull_lights *.py
+recursive-include ap_cull_light *.py
