+++ ap_cull_light/__init__.py
@@ -1,7 +1,7 @@
 """
 Generated By: Cursor (Claude Sonnet 4.5)
 
-ap-cull-lights: A flexible tool for culling astrophotography light frames
+ap-cull-light: A flexible tool for culling astrophotography light frames
 based on FITS header metadata using a rules engine.
 """
