+++ ap_cull_light/__main__.py
@@ -1,7 +1,7 @@
 # Generated By: Claude Code (Claude Sonnet 4.5)
 """Entry point for python -m invocation."""
 
-from ap_cull_lights.cull_lights import main
+from ap_cull_light.cull_lights import main
 
 if __name__ == "__main__":
     main()
