+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "ap-cull-lights"
+name = "ap-cull-light"
 version = "0.1.0"
 description = "Cull astrophotography light frames based on HFR and RMS thresholds"
 readme = "README.md"
@@ -32,7 +32,7 @@ dependencies = [
 ]
 
 [project.scripts]
-ap-cull-lights = "ap_cull_lights.cull_lights:main"
+ap-cull-light = "ap_cull_light.cull_lights:main"
 
 [project.optional-dependencies]
 dev = [
@@ -44,7 +44,7 @@ dev = [
 ]
 
 [tool.setuptools]
-packages = ["ap_cull_lights"]
+packages = ["ap_cull_light"]
 
 [tool.setuptools.package-data]
 "*" = ["*.txt", "*.md"]
