+++ tests/test_cull_lights.py
@@ -1,7 +1,7 @@
 """
 Generated By: Cursor (Claude Sonnet 4.5)
 
-Unit tests for ap_cull_lights.cull_lights module.
+Unit tests for ap_cull_light.cull_lights module.
 """
 
 import os
@@ -10,7 +10,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from ap_cull_lights import cull_lights
+from ap_cull_light import cull_lights
 
 
 class TestRejectImage:
@@ -77,7 +77,7 @@ class TestRejectImage:
         test_file.parent.mkdir(parents=True)
         test_file.write_text("test data")
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             cull_lights.reject_image(
                 filepath=str(test_file),
                 reject_dir=reject_dir,
@@ -126,7 +126,7 @@ class TestRejectImage:
         # Make move_file raise OSError
         mock_move_file.side_effect = OSError("Permission denied")
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             with pytest.raises(OSError):
                 cull_lights.reject_image(
                     filepath=str(test_file),
@@ -154,7 +154,7 @@ class TestRejectImage:
         # Make move_file raise a generic exception
         mock_move_file.side_effect = ValueError("Unexpected error")
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             with pytest.raises(ValueError):
                 cull_lights.reject_image(
                     filepath=str(test_file),
@@ -216,7 +216,7 @@ class TestCullLights:
     """Tests for cull_lights function."""
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_processes_files(
         self, mock_reject, mock_get_metadata, tmp_path
     ):
@@ -261,7 +261,7 @@ class TestCullLights:
         assert mock_reject.called
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_auto_yes_percent(
         self, mock_reject, mock_get_metadata, tmp_path, capsys
     ):
@@ -296,7 +296,7 @@ class TestCullLights:
         assert "y (automatic" in captured.out
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_dryrun(self, mock_reject, mock_get_metadata, tmp_path, capsys):
         """Test that cull_lights works in dryrun mode."""
         source_dir = str(tmp_path / "source")
@@ -326,7 +326,7 @@ class TestCullLights:
         assert "dryrun" in captured.out.lower()
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_skips_files_matching_pattern(
         self, mock_reject, mock_get_metadata, tmp_path
     ):
@@ -364,7 +364,7 @@ class TestCullLights:
         mock_reject.assert_not_called()
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_rms_threshold(self, mock_reject, mock_get_metadata, tmp_path):
         """Test that cull_lights processes RMS threshold correctly."""
         source_dir = str(tmp_path / "source")
@@ -399,7 +399,7 @@ class TestCullLights:
         assert mock_reject.called
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_both_hfr_and_rms(
         self, mock_reject, mock_get_metadata, tmp_path
     ):
@@ -432,7 +432,7 @@ class TestCullLights:
         assert mock_reject.called
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_non_numeric_hfr_debug(
         self, mock_reject, mock_get_metadata, tmp_path, capsys
     ):
@@ -449,7 +449,7 @@ class TestCullLights:
         }
         mock_get_metadata.return_value = mock_metadata
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             cull_lights.cull_lights(
                 source_dir=source_dir,
                 reject_dir=reject_dir,
@@ -465,7 +465,7 @@ class TestCullLights:
             assert "non-numeric HFR" in str(mock_logger.warning.call_args)
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_non_numeric_rms_debug(
         self, mock_reject, mock_get_metadata, tmp_path
     ):
@@ -482,7 +482,7 @@ class TestCullLights:
         }
         mock_get_metadata.return_value = mock_metadata
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             cull_lights.cull_lights(
                 source_dir=source_dir,
                 reject_dir=reject_dir,
@@ -498,7 +498,7 @@ class TestCullLights:
             assert "non-numeric RMS" in str(mock_logger.warning.call_args)
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_missing_filename(
         self, mock_reject, mock_get_metadata, tmp_path
     ):
@@ -517,7 +517,7 @@ class TestCullLights:
         }
         mock_get_metadata.return_value = mock_metadata
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             # Mock input since count_reject is incremented before filename check
             # so a prompt will occur even though rejected_files is empty
             with patch("builtins.input", return_value="y"):
@@ -538,7 +538,7 @@ class TestCullLights:
             mock_reject.assert_not_called()
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_user_rejects(
         self, mock_reject, mock_get_metadata, tmp_path, capsys
     ):
@@ -574,7 +574,7 @@ class TestCullLights:
         assert "Skipping rejection" in captured.out
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_multiple_directory_groups(
         self, mock_reject, mock_get_metadata, tmp_path
     ):
@@ -617,7 +617,7 @@ class TestCullLights:
         assert mock_reject.call_count == 2
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_no_rejections(
         self, mock_reject, mock_get_metadata, tmp_path, capsys
     ):
@@ -652,7 +652,7 @@ class TestCullLights:
         assert "OK to reject" not in captured.out
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_empty_data(
         self, mock_reject, mock_get_metadata, tmp_path, capsys
     ):
@@ -680,7 +680,7 @@ class TestCullLights:
         assert "Done" in captured.out
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_cull_lights.cull_lights.reject_image")
+    @patch("ap_cull_light.cull_lights.reject_image")
     def test_cull_lights_debug_logging(self, mock_reject, mock_get_metadata, tmp_path):
         """Test that cull_lights logs debug information when debug=True."""
         source_dir = str(tmp_path / "source")
@@ -698,7 +698,7 @@ class TestCullLights:
 
         skip_pattern = re.compile("accept")
 
-        with patch("ap_cull_lights.cull_lights.logger") as mock_logger:
+        with patch("ap_cull_light.cull_lights.logger") as mock_logger:
             cull_lights.cull_lights(
                 source_dir=source_dir,
                 reject_dir=reject_dir,
@@ -718,7 +718,7 @@ class TestCullLights:
 class TestMain:
     """Tests for main function."""
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_calls_cull_lights(self, mock_cull, tmp_path):
         """Test that main function calls cull_lights with correct arguments."""
         source_dir = str(tmp_path / "source")
@@ -738,7 +738,7 @@ class TestMain:
         assert call_kwargs["max_hfr"] == 2.5
         assert call_kwargs["max_rms"] is None
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_with_max_rms(self, mock_cull, tmp_path):
         """Test main with --max-rms parameter."""
         source_dir = str(tmp_path / "source")
@@ -755,7 +755,7 @@ class TestMain:
         assert call_kwargs["max_hfr"] is None
         assert call_kwargs["max_rms"] == 2.0
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_with_both_params(self, mock_cull, tmp_path):
         """Test main with both --max-hfr and --max-rms."""
         source_dir = str(tmp_path / "source")
@@ -780,7 +780,7 @@ class TestMain:
         assert call_kwargs["max_hfr"] == 2.5
         assert call_kwargs["max_rms"] == 2.0
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_with_debug(self, mock_cull, tmp_path):
         """Test main with --debug flag."""
         source_dir = str(tmp_path / "source")
@@ -796,7 +796,7 @@ class TestMain:
         call_kwargs = mock_cull.call_args[1]
         assert call_kwargs["debug"] is True
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_with_dryrun(self, mock_cull, tmp_path):
         """Test main with --dryrun flag."""
         source_dir = str(tmp_path / "source")
@@ -822,7 +822,7 @@ class TestMain:
             with pytest.raises(SystemExit):
                 cull_lights.main()
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_with_skip_regex(self, mock_cull, tmp_path):
         """Test main with --skip-regex parameter."""
         source_dir = str(tmp_path / "source")
@@ -847,7 +847,7 @@ class TestMain:
         assert call_kwargs["skip_pattern"] is not None
         assert call_kwargs["skip_pattern"].pattern == "accept"
 
-    @patch("ap_cull_lights.cull_lights.cull_lights")
+    @patch("ap_cull_light.cull_lights.cull_lights")
     def test_main_with_auto_accept_percent(self, mock_cull, tmp_path):
         """Test main with --auto-accept-percent parameter."""
         source_dir = str(tmp_path / "source")
